student_tagger_app/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
jinja2>=3.1,<4.0
python-multipart>=0.0.9,<1.0
Pillow>=10.0,<12.0
numpy>=1.26,<3.0
dlib>=19.24
face_recognition_models
